chore: refactor install workflows to be centralized
