+++ services/ctl-api/internal/app/installs/helpers/create_install_workflow.go
@@ -6,7 +6,9 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/nuonco/nuon/services/ctl-api/internal/app"
+	"github.com/nuonco/nuon/services/ctl-api/internal/app/installs/signals/v2/generateworkflowsteps"
 	"github.com/nuonco/nuon/services/ctl-api/internal/pkg/db/generics"
+	signaldb "github.com/nuonco/nuon/services/ctl-api/internal/pkg/queue/signal/db"
 )
 
 func (s *Helpers) CreateWorkflowWithRole(
@@ -77,6 +79,9 @@ func (s *Helpers) createWorkflow(ctx context.Context,
 		ApprovalOption:    approvalOption,
 		PlanOnly:          planOnly,
 		Role:              role,
+		GenerateStepsSignal: &signaldb.SignalData{
+			Signal: &generateworkflowsteps.Signal{},
+		},
 	}
 
 	res := s.db.WithContext(ctx).Create(&installWorkflow)
