+++ services/ctl-api/internal/app/installs/service/retry_install_workflow.go
@@ -9,7 +9,6 @@ import (
 	"github.com/nuonco/nuon/services/ctl-api/internal/app"
 	"github.com/nuonco/nuon/services/ctl-api/internal/app/installs/helpers"
 	"github.com/nuonco/nuon/services/ctl-api/internal/app/installs/signals"
-	rerunflow "github.com/nuonco/nuon/services/ctl-api/internal/app/installs/signals/v2/rerunflow"
 	"github.com/nuonco/nuon/services/ctl-api/internal/middlewares/stderr"
 	"github.com/nuonco/nuon/services/ctl-api/internal/pkg/cctx"
 	validatorPkg "github.com/nuonco/nuon/services/ctl-api/internal/pkg/validator"
@@ -135,38 +134,16 @@ func (s *service) RetryWorkflow(ctx *gin.Context) {
 		}
 	}
 
-	useQueues, err := s.featuresClient.AllFeaturesEnabled(ctx, app.OrgFeatureAppBranches, app.OrgFeatureQueues)
-	if err != nil {
-		ctx.Error(fmt.Errorf("checking features: %w", err))
-		return
-	}
-	if useQueues {
-		queueID, err := s.getInstallWorkflowsQueueID(ctx, install_id)
-		if err != nil {
-			ctx.Error(err)
-			return
-		}
-		if err := s.enqueueInstallSignal(ctx, queueID, &rerunflow.Signal{
-			InstallID:         install_id,
-			InstallWorkflowID: workflow.ID,
-			RerunConfiguration: signals.RerunConfiguration{
-				StepID:        req.StepID,
-				StepOperation: signals.RerunOperation(req.Operation),
-			},
-		}, workflow.ID, "install_workflows"); err != nil {
-			ctx.Error(fmt.Errorf("enqueue signal: %w", err))
-			return
-		}
-	} else {
-		s.evClient.Send(ctx, install_id, &signals.Signal{
-			Type:              signals.OperationRerunFlow,
-			InstallWorkflowID: workflow.ID,
-			RerunConfiguration: signals.RerunConfiguration{
-				StepID:        req.StepID,
-				StepOperation: signals.RerunOperation(req.Operation),
-			},
-		})
-	}
+	// NOTE: rerunflow signal has been removed. This deprecated endpoint
+	// uses the legacy event loop path only.
+	s.evClient.Send(ctx, install_id, &signals.Signal{
+		Type:              signals.OperationRerunFlow,
+		InstallWorkflowID: workflow.ID,
+		RerunConfiguration: signals.RerunConfiguration{
+			StepID:        req.StepID,
+			StepOperation: signals.RerunOperation(req.Operation),
+		},
+	})
 
 	ctx.JSON(201, RetryWorkflowResponse{
 		WorkflowID: workflow.ID,
