+++ services/ctl-api/internal/fxmodules/workers_shared.go
@@ -33,7 +33,8 @@ import (
 	onboardingactivities "github.com/nuonco/nuon/services/ctl-api/internal/app/onboarding/signals/activities"
 
 	// Register install queue signals
-	_ "github.com/nuonco/nuon/services/ctl-api/internal/app/installs/signals/v2/executeflow"
+	_ "github.com/nuonco/nuon/services/ctl-api/internal/app/installs/signals/v2/generateworkflowsteps"
+	_ "github.com/nuonco/nuon/services/ctl-api/internal/pkg/flow/signals/executeflow"
 
 	// Register VCS queue signals
 	_ "github.com/nuonco/nuon/services/ctl-api/internal/app/vcs/signals/v2/healthcheck"
