+++ services/ctl-api/internal/pkg/flow/client/approve_plan.go
@@ -7,7 +7,7 @@ import (
 	tclient "go.temporal.io/sdk/client"
 
 	"github.com/nuonco/nuon/services/ctl-api/internal/app"
-	"github.com/nuonco/nuon/services/ctl-api/internal/app/installs/signals/v2/executeflow"
+	"github.com/nuonco/nuon/services/ctl-api/internal/pkg/flow/signals/executeflow"
 )
 
 // ApprovePlanRequest is the input for approving a plan on a workflow step.
